requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
requests==2.31.0
PyGithub==2.1.1
openpyxl==3.1.2
numpy==1.24.3
pandas==2.0.3
xlrd==2.0.1
